Add POC for data-preserve-html attribute
Allow divs to have the data-preserve-html attribute, which does two
things:
 1) Stops the insertion of p tags
 2) Prevents it from being editable in the Rich Text view

The p tags are inserted to help make the Rich Text view functional, as
otherwise it is difficult and error-prone to manipulate text when there
is no text element at that position in the hierarchy.

However, for some work, especially those that use skins to make the
content look like another existing service, the insertion of the extra
tags makes styling difficult and unpredictable.

The data-preserve-html attribute would let users opt-out of the
insertion of extra tags, at the cost of no longer being able to edit
that part of the content in the Rich Text view.

To achieve this we:
 1) Add data-preserve-html to the allowlist for divs
 2) Skip the p tag insertion if the tag has the data-preserve-html
 attribute set to "true"
 3) Add a custom plugin to mce to make any element with the
 data-preserve-html attribute set to true uneditable. This is also
 accompanied by a message explaining why.